Add files via upload
Anato is a man who has been notified there is a zombie attack in the town. So, he has to get to a safe bunker(there are multiple). On his way , he can meet zombie(Z) who would kill him but if he obtains a Weapon(W) he can kill as many zombies as the number of W he got. And so I have implemented the BFS,UCS,A* and DFS to get him to the bunkers.